Refatora aba de licenças em componente dedicado
